+++ server/impl/src/main/java/com/walmartlabs/concord/server/audit/AuditObject.java
@@ -29,5 +29,6 @@ public enum AuditObject {
     PROJECT,
     SECRET,
     SYSTEM,
-    TEAM
+    TEAM,
+    ROLE
 }