+++ server/impl/src/test/java/com/walmartlabs/concord/server/user/UserDaoTest.java
@@ -49,7 +49,7 @@ public class UserDaoTest extends AbstractDaoTest {
     public void testInsertDelete() throws Exception {
         String username = "user#" + System.currentTimeMillis();
 
-        UUID userId = userDao.insert(username, UserType.LOCAL);
+        UUID userId = userDao.insert(username, UserType.LOCAL, null);
 
         String s = "key#" + System.currentTimeMillis();
         String name = "name#" + System.currentTimeMillis();